requirements.txt
Flask==2.3.2
numpy==1.24.3
opencv_python==4.7.0.72
paho_mqtt==1.6.1
PyYAML==6.0
tflite_support==0.4.3
# Note: If this doesn't work, you might need to install the package using pip directly:
# pip install --extra-index-url https://google-coral.github.io/py-repo/ tflite_support
requests==2.30.0
Pillow==9.5.0
watchdog==3.0.0